1.8/NewHelpNotary/requirements.txt
pillow
django-admin-interface
django-crispy-formspip 